Added the functions demonstrating the basic framework of the game
To make the code easier to colloborate on and for simplicities sake, our game will use a state based control system, with various descriptive strings representing various gameStates. These game states will be used as keys in objects that will go in essential game functions (draw function, mousePressed, ect.) that  contian references to corresponding functions. In this commit we've added mousePressed and draw functions for the "startScreen" and "inGame" states.